+++ src/Support/GroupSessionConfigException.php
@@ -1,0 +1,9 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Ometra\Caronte\Support;
+
+use RuntimeException;
+
+final class GroupSessionConfigException extends RuntimeException {}